% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wareki2.R
\name{wareki2}
\alias{wareki2}
\title{Convert Japanese Calendar, Gen-gou, into Westerm Year in sentences.}
\usage{
wareki2(textdata, dummyText = NA, loop = FALSE)
}
\arguments{
\item{textdata}{text data including Japanese Calendar.}

\item{dummyText}{If something trouble with the output, the dummyText, a piecs of text temporary added to document, may be the cause. If needed, instead of the default dummyText, you may use your original dummyText, which is recommended to be a reasonably long String without using numeric.}

\item{loop}{Whether the process applies again and again to a part once converted. Default FALSE is strongly recommended.}
}
\value{
document
}
\description{
Convert Japanese Calendar, Gen-gou, into Westerm Year in sentences.
}
